Doc : technologies part done
Tag Engine : socket unix server that respond to queries for listing
existent tags and list of files associated to an one tag only (for now).
Construct the full path of files.

Tag Manager : socket unix client, two new options to get existent tags
and files associated to an one tag only (for now).

On branch master

Changes to be committed:
	modified:   doc
	modified:   makefile
	modified:   tag_engine
	modified:   tag_manager